Refactor: Merge save/load into one file, remove verbose from MCP context
