--- e2etests/tests/test_save_command_argument_validation.rs
@@ -1,46 +1,0 @@
-use q_cli_e2e_tests::q_chat_helper::QChatSession;
-
-#[test]
-#[cfg(feature = "save_load")]
-fn test_save_command_argument_validation() -> Result<(), Box<dyn std::error::Error>> {
-    println!("🔍 Testing /save command...");
-    
-    let mut chat = QChatSession::new()?;
-    println!("✅ Q Chat session started");
-    
-    let response = chat.execute_command("/save")?;
-    
-    println!("📝 Help response: {} bytes", response.len());
-    println!("📝 FULL OUTPUT:");
-    println!("{}", response);
-    println!("📝 END OUTPUT");
-    
-    // Verify save error message
-    assert!(response.contains("error:") && response.contains("the following required arguments were not provided:"), "Missing save error message");
-    println!("✅ Found save error message");
-    
-    assert!(response.contains("Usage:"), "Missing Usage section");
-    assert!(response.contains("/save"), "Missing /save command in usage");
-    println!("✅ Found Usage section with /save command");
-    
-    assert!(response.contains("Arguments:"), "Missing Arguments section");
-    assert!(response.contains("<PATH>"), "Missing PATH argument");
-    println!("✅ Found Arguments section with PATH parameter");
-    
-    assert!(response.contains("Options:"), "Missing Options section");
-    assert!(response.contains("-f"), "Missing -f flag");
-    assert!(response.contains("--force"), "Missing --force flag");
-    assert!(response.contains("-h"), "Missing -h flag");
-    assert!(response.contains("--help") || response.contains("—help"), "Missing --help flag");
-    println!("✅ Found Options section with -f, --force, -h, --help flags");
-    
-    assert!(response.contains("Print help"), "Missing help description");
-    println!("✅ Found help flag description");
-    
-    println!("✅ All help content verified!");
-    
-    chat.quit()?;
-    println!("✅ Test completed successfully");
-    
-    Ok(())
-}
